refactor(playground/web): remove deprecated template-chat-utils and introduce new utility files for JSON patch handling and date formatting
